+++ src/Utility/FileUtils.hpp
@@ -18,11 +18,6 @@ namespace fs = std::experimental::filesystem;
 #endif
 #include "FinalAction.hpp"
 
-#if defined(ENIGMA_PLATFORM_WINDOWS)
-#include <fcntl.h>
-#include <io.h>
-#endif
-
 NS_ENIGMA_BEGIN
 class FileUtils final {
     ENIGMA_STATIC_CLASS(FileUtils);
@@ -113,9 +108,6 @@ class FileUtils final {
         ENIGMA_ERROR("Failed to read file chunks {}", filename.string());
         return;
       }
-#if defined(ENIGMA_PLATFORM_WINDOWS)
-      _setmode(_fileno(file), _O_BINARY);
-#endif
       FinalAction fileCloser([&file] {
         std::fclose(file);
       });